requirements.txt
# Core dependencies for OpenAI pipeline
openai==1.12.0
python-dotenv==1.0.0
Pillow==10.2.0
numpy>=1.26.0  # Compatible with Python 3.12
httpx  # For proxy handling
requests  # For downloading images

# Utilities
tqdm==4.66.1

# Optional background removal (if needed)
# rembg==2.0.56

# Optional for future enhancements
# streamlit==1.31.0  # For web interface
# opencv-python==4.9.0.80  # For image quality checks

# Development tools (optional)
# pytest==8.0.0
# black==24.1.1
# flake8==7.0.0 